% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Categorical Data tests.R
\name{ChiSquareTestHomogeneity}
\alias{ChiSquareTestHomogeneity}
\title{Chi-Square test for Homogeneity}
\usage{
ChiSquareTestHomogeneity(data, alpha)
}
\arguments{
\item{data}{Contingency table dataframe!!!}

\item{alpha}{Significance level}
}
\value{
A list with Marginals, expected values, chi-s, degrees of freedom, p value and if H0 is TRUE or FALSE
}
\description{
Hyphotesis test for multiple Multinomial distribution (k=number of outcomes). The null hypothesis is that
the probabilities of the outcomes are the same for each experiment. Chi-square dist. Allways look to the right hand tail.\cr
H0: for each column j, p1j=..=pIj
}
\details{
Import data frame as:\cr
             less1 X1to5 graterthan5\cr
Diseased      10     8          23\cr
Sensitized     9    19          11\cr
Normal        70   136         206\cr

EXPECTED values should be >5. Df=(I(rows)-1)(J(col)-1).
}
