% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Differences between two Proportions.R
\name{DifferenceBetweenProportions}
\alias{DifferenceBetweenProportions}
\title{Test for the Differences between two Proportions}
\usage{
DifferenceBetweenProportions(nx, x, ny, y, type, alpha)
}
\arguments{
\item{x}{Number of successes of X}

\item{y}{Number of successes of Y}

\item{type}{is character "R" right tail ,"L" left or "C" centered}

\item{alpha}{Significance value}

\item{n}{x,y sample sizes}
}
\value{
Data frame with if H0 is TRUE or FALSE, pvalue and z-score
}
\description{
Hyphotesis test for the Differences between two Proportions Let X ∼ Bin(nX, pX) and let Y ∼ Bin(nY, pY).
Assume that there are at least 10 successes and 10 failures in each sample, and that X and Y are "INDEPENDENT".
}
\details{
H1:px-py>0   H0:px-py<=0 ---->  "R"\cr
H1:px-py<0   H0:px-py>=0 ---->  "L"\cr
H1:px-py!=0  H0:px-py= 0  ---->  "C"\cr
}
