% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test for Variances of Normal Popus.R
\name{FTestForEqualityOfVariances}
\alias{FTestForEqualityOfVariances}
\title{F-Test for equality of variance for NOrmal Pop.}
\usage{
FTestForEqualityOfVariances(x, y, type)
}
\arguments{
\item{x}{Vector Values from aNorm Dist}

\item{y}{Vector Values from aNorm Dist}

\item{type}{is character "R" right tail ,"L" left or "C" centered}
}
\value{
Data frame with if H0 is TRUE or FALSE, pvalue and z-score
}
\description{
Hyphotesis test for a normal population concerning variances. Null distribution is F: sx^2/sy^2
}
\details{
H0:σx^2/σy^2<=1 ---->  "R"\cr
H0:σx^2/σy^2=>1 ---->  "L"\cr
H0:σx^2/σy^2!=1  ---->  "C"\cr
}
