% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Categorical Data tests.R
\name{GoodnessOfFitTest}
\alias{GoodnessOfFitTest}
\title{Goodness-Of-Fit Test}
\usage{
GoodnessOfFitTest(x, alpha)
}
\arguments{
\item{x}{Vector with the observed values of the Categories}
}
\value{
A list with K, expected values vector, chi-s, degrees of freedom, p value and if H0 is TRUE or FALSE
}
\description{
Hyphotesis test for Multinomial distribution (k=number of outcomes). If H0 is TRUE the observation and Expected values
are very likely close to each other. Chi-square dist. Allways look to the right hand tail.
}
\details{
EXPECTED values should be >5. Df=k-1.

H0: p1=p01,p2=p02,p3=p03,..,pk=p0k
}
