% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Large-Sample test for Mean Differences.R
\name{LargeSampleTestMeanDiff}
\alias{LargeSampleTestMeanDiff}
\title{Large-Sample Test for mean differences}
\usage{
LargeSampleTestMeanDiff(nx, ux, sx, ny, uy, sy, d0, type, alpha)
}
\arguments{
\item{ux}{mean of the sample x}

\item{sx}{sd fo the sample X}

\item{uy}{mean of the sample y}

\item{sy}{sd fo the sample Y}

\item{d0}{Testing value}

\item{type}{is character "R" right tail ,"L" left or "C" centered}

\item{alpha}{significance value}

\item{n}{x,y sample sizes}
}
\value{
Data frame with if H0 is TRUE or FALSE, pvalue and z-score
}
\description{
Hyphotesis test for Large-Sample Test for mean differences which n is large (e.g., nX > 30 and nY > 30)
}
\details{
ASSUME the samples are drawn INDEPENDENTLY of each other.

H1:ux-uy>d0   H0:ux-uy<=d0 ---->  "R"\cr
H1:ux-uy<d0   H0:ux-uy>=d0 ---->  "L"\cr
H1:ux-uy!=d0  H0:ux-uy=d0  ---->  "C"\cr
}
