% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyphotesisTestingDs.R
\name{LargeTest}
\alias{LargeTest}
\title{Large sample test}
\usage{
LargeTest(n, ux, sx, u0, type, alpha)
}
\arguments{
\item{n}{sample size}

\item{ux}{mean of the sample}

\item{sx}{standard deviation of the sample}

\item{u0}{hyphotesis value}

\item{type}{is character "R","L"or"C"}

\item{alpha}{significance value}
}
\value{
Data frame with if H0 is TRUE or FALSE, pvalue and z-score
}
\description{
Hyphotesis test for large samples: n>30
}
