% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test with paired data.R
\name{PairedDataTest}
\alias{PairedDataTest}
\title{Test with paired data}
\usage{
PairedDataTest(n, uD, sD, u0, type, alpha)
}
\arguments{
\item{n}{sample size}

\item{uD}{mean of the DIFFERENCES between samples}

\item{sD}{sd fo the DIFFERENCES between samples}

\item{u0}{Testing value}

\item{type}{is character "R" right tail ,"L" left or "C" centered}

\item{alpha}{significance value}
}
\value{
Data frame with if H0 is TRUE or FALSE, pvalue and z-score
}
\description{
Hyphotesis test for Paired Data. CHECK if there is any relation between variables e,g. Vehicle, and Tires A,B.
The samples are not INDEPENDET.Let (X1, Y1), …, (Xn, Yn) be a sample of ordered pairs whose differences D1,...,Dn are a sample
from a normal population with mean μD. Let sD be the sample standard deviation of D1, …,Dn.
}
\details{
The the statistic test is Student's t with n-1 degrees of freedom. If the sample is LARGE use z instead of t

H1:uD>u0   H0:uD<=u0 ---->  "R"\cr
H1:uD<u0   H0:uD>=u0 ---->  "L"\cr
H1:uD!=u0  H0:uD=u0  ---->  "C"\cr
}
