% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyphotesisTestingDs.R
\name{ProportionTest}
\alias{ProportionTest}
\title{Proportion test}
\usage{
ProportionTest(n, px, p0, type, alpha)
}
\arguments{
\item{n}{sample size}

\item{px}{of the sample}

\item{p0}{hyphotesis value}

\item{type}{is character "R" right tail ,"L" left or "C" centered}

\item{alpha}{significance value}
}
\value{
Data frame with if H0 is TRUE or FALSE, pvalue and z-score
}
\description{
Hyphotesis test for a population proportion n*p0 and n(1-p0)>10
}
\details{
Let X be the number of succeses in "n" independent Bernoulli trials with probability p: X~Bin(n,p)\cr

H1:p>p0   H0:p<=p0 "R"\cr
H1:p<p0   H0:p>=p0 "L"\cr
H1:p!=p0  H0:p=p0  "C"\cr
}
