% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Small-Sample Mean Differences.R
\name{SmallSampleMeanDifferences}
\alias{SmallSampleMeanDifferences}
\title{Small-Sample Test for the difference between two means}
\usage{
SmallSampleMeanDifferences(nx, ux, sx, ny, uy, sy, d0, type, alpha)
}
\arguments{
\item{ux}{mean of the sample X}

\item{sx}{sd fo the sample X}

\item{uy}{mean of the sample Y}

\item{sy}{sd fo the sample Y}

\item{d0}{Testing value}

\item{type}{is character "R" right tail ,"L" left or "C" centered}

\item{alpha}{significance value}

\item{n}{x,y sample size of X and Y respectively}
}
\value{
Data frame with if H0 is TRUE or FALSE, pvalue and z-score
}
\description{
Let X1,...,Xnx and Y1,...,Ynx be samples from normal populations with means μX and μY and standard deviations σX and σY, respectively.
Assume the samples are drawn "independently" of each other.
}
\details{
H1:ux-uy>d0   H0:ux-uy<=d0 ---->  "R"\cr
H1:ux-uy<d0   H0:ux-uy>=d0 ---->  "L"\cr
H1:ux-uy!=d0  H0:ux-uy=d0  ---->  "C"\cr
}
