% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Small-Sample Population Mean.R
\name{SmallSampleMeanTest}
\alias{SmallSampleMeanTest}
\title{Small Sample Test for a population Mean}
\usage{
SmallSampleMeanTest(n, ux, sx, u0, type, alpha)
}
\arguments{
\item{n}{sample size}

\item{ux}{mean of the sample}

\item{sx}{sd fo the sample}

\item{u0}{Testing value}

\item{type}{is character "R" right tail ,"L" left or "C" centered}

\item{alpha}{significance value}
}
\value{
Data frame with if H0 is TRUE or FALSE, pvalue and z-score
}
\description{
Hyphotesis test for a Small Sample Test for a population Mean n<30 that comes from a NORMAL dist. sigma is unknown
}
\details{
Let X be the number of succeses in "n" independent Bernoulli trials with probability p: X~Bin(n,p)\cr

H1:p>p0   H0:p<=p0 ---->  "R"\cr
H1:p<p0   H0:p>=p0 ---->  "L"\cr
H1:p!=p0  H0:p=p0  ---->  "C"\cr
}
