% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test for Variances of Normal Popus.R
\name{TestVarianceOfNormPop}
\alias{TestVarianceOfNormPop}
\title{Test for Variances of Normal Populations}
\usage{
TestVarianceOfNormPop(n, s2x, s20, type, alpha)
}
\arguments{
\item{n}{sample size}

\item{s2x}{Variance for the sample}

\item{s20}{Testing variance value}

\item{type}{is character "R" right tail ,"L" left or "C" centered}

\item{alpha}{significance value}
}
\value{
Data frame with if H0 is TRUE or FALSE, pvalue and z-score
}
\description{
Hyphotesis test for a normal population concerning variances.
}
\details{
Let X1, …,Xn be a simple random sample from a N(μ, σ2) population. Let s^2 be the sample variance:\cr

H1:σ^2>σ0^2   H0:σ^2<=σ0^2 ---->  "R"\cr
H1:σ^2<σ0^2   H0:σ^2>=σ0^2 ---->  "L"\cr
H1:σ^2!=σ0^2  H0:σ^2=σ0^2  ---->  "C"\cr
}
